Updates for the "teams toolbar" - refactored component to a common structure - fixed test
