+++ app/teams/(teams-page)/@content/TeamsContent.tsx
@@ -5,7 +5,7 @@ import { ITeamsSearchParams } from '@/types/teams.types';
 import { IUserInfo } from '@/types/shared.types';
 import EmptyResult from '../../../../components/core/empty-result';
 import Error from '../../../../components/core/error';
-import TeamsToolbar from '../../../../components/page/teams/teams-toolbar';
+import { TeamsToolbar } from '../../../../components/page/teams/TeamsToolbar';
 import TeamList from '@/components/page/teams/team-list';
 import styles from './page.module.css';
 import { fetchTeamsList } from '../teamsApi';
