+++ components/page/teams/TeamsToolbar/TeamsToolbar.module.scss
@@ -1,0 +1,141 @@
+.toolbar {
+  display: flex;
+  justify-content: space-between;
+  height: 40px;
+  gap: 8px;
+  width: 100%;
+}
+
+.left {
+  display: flex;
+  gap: 8px;
+}
+
+.filterBtn {
+  padding: 8px 12px;
+  border: none;
+  display: flex;
+  gap: 8px;
+  align-items: center;
+  border-radius: 8px;
+  background: #fff;
+  height: inherit;
+  box-shadow: 0px 1px 2px 0px rgba(15, 23, 42, 0.16);
+}
+
+.titleContainer {
+  display: none;
+}
+
+.searchContainer {
+  height: inherit;
+  border-radius: 4px;
+  background: #fff;
+  box-shadow: 0px 1px 2px 0px rgba(15, 23, 42, 0.16);
+  display: none;
+}
+
+.searchForm {
+  display: flex;
+  height: 40px;
+  padding: 8px 12px;
+  border-radius: 4px;
+  align-items: center;
+  gap: 10px;
+}
+
+.searchOptions {
+  display: flex;
+  width: 40px;
+  justify-content: end;
+  position: relative;
+  align-items: center;
+  gap: 10px;
+}
+
+.clearBtn {
+  background: inherit;
+  outline: none;
+  position: absolute;
+  bottom: 0;
+  height: 16px;
+  left: 0;
+}
+
+.searchInput {
+  border: none;
+  width: 84px;
+  background-color: #fff;
+  color: black;
+
+  &:focus {
+    outline: none;
+  }
+
+  &::placeholder {
+    color: #94a3b8;
+    font-size: 14px;
+    font-weight: 500;
+    line-height: 24px;
+  }
+}
+
+.searchBtn {
+  border: none;
+  background: #fff;
+  height: 16px;
+}
+
+.right {
+  display: flex;
+  gap: 8px;
+  align-items: center;
+}
+
+@media (min-width: 480px) {
+  .searchInput {
+    width: 150px;
+  }
+}
+
+@media (min-width: 1024px) {
+  .toolbar {
+    background-color: #f1f5f9;
+  }
+
+  .left {
+    gap: 16px;
+  }
+
+  .filterBtn {
+    display: none;
+  }
+
+  .titleContainer {
+    display: flex;
+    gap: 8px;
+    align-items: baseline;
+  }
+
+  .searchInput {
+    width: 192px;
+  }
+
+  .title {
+    color: #0f172a;
+    font-size: 28px;
+    font-weight: 700;
+    line-height: 40px;
+  }
+
+  .count {
+    color: #64748b;
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 20px;
+  }
+
+  .right {
+    gap: 16px;
+  }
+}
