+++ components/page/teams/TeamsToolbar/index.ts
@@ -1,0 +1,1 @@
+export * from './TeamsToolbar';
